requirements/common.txt
django==5.0.3
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
black==24.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.6
Django==5.0.3
django-binary-database-files==1.0.18
django-filter==24.2
djangorestframework==3.15.1
fastapi==0.103.1
fenics==2019.1.0
fenics-dijitso==2019.1.0
fenics-ffc==2019.1.0.post0
fenics-fiat==2019.1.0
fenics-ufl==2019.1.0
filelock==3.10.0
flake8==7.0.0
fonttools==4.40.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
numpy==1.24.2
packaging==23.1
pandas==1.5.3
pathspec==0.12.1
Pillow==9.5.0
platformdirs==3.1.1
plotly==5.15.0
protobuf==4.21.12
pycodestyle==2.11.1
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0.1
readchar==4.0.5
rich==13.7.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2024.1
uvicorn==0.23.2
vboxapi==1.0
virtualenv==20.21.0
watchfiles==0.20.0
websockets==11.0.3